+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## HEAD
+
 ## v1.1.0
 
 * Fix example for JSONP response type (typo: `\` instead of `"`)